requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
elevenlabs==2.23.0
eval_type_backport==0.3.0
fastapi==0.121.2
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.3.0
invoke==2.2.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistralai==1.9.11
multidict==6.7.0
narwhals==2.11.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==12.0.0
pluggy==1.6.0
portalocker==3.2.0
postgrest==2.25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyogrio==0.10.0
pyproj==3.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
realtime==2.25.0
redis==7.1.0
referencing==0.37.0
requests==2.32.3
rpds-py==0.28.0
rsa==4.9.1
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.3
storage3==2.25.0
streamlit==1.51.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
